rekruting/requirements.txt
aiogram==3.21.0
SQLAlchemy>=2.0,<2.1
aiosqlite>=0.20.0
python-dotenv>=1.0.1
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
